+++ abci/code/code.go
@@ -100,5 +100,8 @@ const (
 	NodeIDisProxyNode                         uint32 = 73
 	NodeIDHasNotBeenAssociatedWithProxyNode   uint32 = 74
 	ProxyNodeNotFound                         uint32 = 75
+	NodeIDIsNotExistInIdPList                 uint32 = 76
+	ProxyNodeIsNotActive                      uint32 = 77
+	NodeIDInIdPListIsNotActive                uint32 = 78
 	UnknownError                              uint32 = 999
 )